Code for suicidal bullet done.

+++ Astro Party/Assets/Yuxiang/Scripts/BulletMove.cs
@@ -11,6 +11,7 @@ public class BulletMove : MonoBehaviour
     int speed = 1000;
 
     GameManager gameManagerScript;
+    ScoreManager scoreManagerScript;
 
     bool attacked;
 
@@ -20,6 +21,7 @@ public class BulletMove : MonoBehaviour
         Rb = GetComponent<Rigidbody>();
         Rb.AddRelativeForce(new Vector3(0, 0, speed), ForceMode.VelocityChange);
         gameManagerScript = GameObject.Find("Game Manager").GetComponent<GameManager>();
+        scoreManagerScript = GameObject.Find("Score Manager").GetComponent<ScoreManager>();
     }
 
     // Update is called once per frame
@@ -46,17 +48,46 @@ public class BulletMove : MonoBehaviour
             {
                 if (collision.gameObject.GetComponent<PilotPlayerController>() != null)
                 {
-                    collision.gameObject.GetComponent<PilotPlayerController>().kill(id, team);
+                    if (collision.gameObject.GetComponent<PilotPlayerController>().id == id)
+                    {
+                        if (gameManagerScript.suicidalBullet)
+                        {
+                            suicide();
+                        }
+                    }
+                    else
+                    {
+                        collision.gameObject.GetComponent<PilotPlayerController>().kill(id, team);
+                    }
                 }
                 else
                 {
-                    collision.gameObject.GetComponent<BotPilotMove>().kill(id, team);
+                    if (collision.gameObject.GetComponent<BotPilotMove>().id == id)
+                    {
+                        if (gameManagerScript.suicidalBullet)
+                        {
+                            suicide();
+                        }
+                    }
+                    else
+                    {
+                        collision.gameObject.GetComponent<BotPilotMove>().kill(id, team);
+                    }
                 }
             }
-
             if (collision.gameObject.CompareTag("Ship"))
             {
-                collision.gameObject.GetComponent<MutualShip>().damage(id, team);
+                if (collision.gameObject.GetComponent<MutualShip>().id == id)
+                {
+                    if (gameManagerScript.suicidalBullet)
+                    {
+                        suicide();
+                    }
+                }
+                else
+                {
+                    collision.gameObject.GetComponent<MutualShip>().damage(id, team);
+                }
             }
 
             if (!collision.gameObject.CompareTag("Floor"))
@@ -64,11 +95,6 @@ public class BulletMove : MonoBehaviour
                 bool destroy = true;
                 if (collision.gameObject.CompareTag("Bullet"))
                 {
-                    if (gameManagerScript.bulletCancel)
-                    {
-                        destroy = false;
-                    }
-
                     //for scatter shot
                     if (collision.gameObject.GetComponent<BulletMove>().id == id)
                     {
@@ -81,6 +107,31 @@ public class BulletMove : MonoBehaviour
                     Destroy(gameObject);
                     attacked = true;
                 }
+            }
+        }
+    }
+
+    void suicide()
+    {
+        if (scoreManagerScript.gameMode == "solo")
+        {
+            switch (id)
+            {
+                case 1:
+                    scoreManagerScript.P1Suicide = true;
+                    break;
+                case 2:
+                    scoreManagerScript.P2Suicide = true;
+                    break;
+                case 3:
+                    scoreManagerScript.P3Suicide = true;
+                    break;
+                case 4:
+                    scoreManagerScript.P4Suicide = true;
+                    break;
+                case 5:
+                    scoreManagerScript.P5Suicide = true;
+                    break;
             }
         }
     }
